web_manage/requirements.txt
absl-py==0.8.1
amqp==1.4.9
anyjson==0.3.3
astor==0.8.0
autopep8==1.4.4
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
Django==2.2.3
django-celery==3.3.1
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-redis==4.10.0
DjangoUeditor==1.8.143
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kombu==3.0.37
Markdown==3.1.1
meld3==2.0.0
numpy==1.17.2
opt-einsum==3.1.0
pandas==0.25.1
Pillow==6.1.0
protobuf==3.10.0
pycodestyle==2.5.0
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
redis==2.10.6
six==1.12.0
sqlparse==0.3.0
supervisor==4.0.4
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
virtualenv==16.6.2
Werkzeug==0.16.0
wrapt==1.11.2
xlrd==1.2.0
